Replica API sugar
