+++ src/test/replica/query_source.test.ts
@@ -28,11 +28,7 @@ Deno.test("QuerySource", async () => {
         docDriver: new DocDriverMemory(SHARE_ADDR),
         attachmentDriver: new AttachmentDriverMemory(),
       },
-      config: {
-        "es.5": {
-          shareSecret: shareKeypair.secret,
-        },
-      },
+      shareSecret: shareKeypair.secret,
     },
   );
