+++ src/test/replica/replica_config.test.ts
@@ -59,11 +59,6 @@ let _runStorageConfigTests = (
           docDriver: driver,
           attachmentDriver: new AttachmentDriverMemory(),
         },
-        config: {
-          "es.5": {
-            shareSecret: "",
-          },
-        },
       })
       : driver;
   };
