+++ src/test/syncer/have_entry_keeper.test.ts
@@ -32,9 +32,7 @@ Deno.test("HaveEntryKeeper", async () => {
         docDriver: new DocDriverMemory(SHARE_ADDR),
         attachmentDriver: new AttachmentDriverMemory(),
       },
-      config: {
-        "es.5": { shareSecret: shareKeypair.secret },
-      },
+      shareSecret: shareKeypair.secret,
     },
   );
 
@@ -185,11 +183,7 @@ Deno.test({
         docDriver: new DocDriverMemory(SHARE_ADDR),
         attachmentDriver: new AttachmentDriverMemory(),
       },
-      config: {
-        "es.5": {
-          shareSecret: shareKeypair.secret,
-        },
-      },
+      shareSecret: shareKeypair.secret,
     });
 
     await writeRandomDocs(keypairA, replica, 1000);
@@ -199,11 +193,7 @@ Deno.test({
         docDriver: new DocDriverMemory(SHARE_ADDR),
         attachmentDriver: new AttachmentDriverMemory(),
       },
-      config: {
-        "es.5": {
-          shareSecret: shareKeypair.secret,
-        },
-      },
+      shareSecret: shareKeypair.secret,
     });
 
     const ingestWritable = new WritableStream<QuerySourceEvent<DocEs5>>({
